requirements.txt
#

# These requirements were autogenerated by pipenv

# To regenerate from the project's Pipfile, run:

#

#    pipenv lock --requirements

#



-i https://pypi.org/simple

cachetools==5.0.0; python_version ~= '3.7'

certifi==2021.10.8

charset-normalizer==2.0.12; python_version >= '3'

click==8.1.3; python_version >= '3.7'

colorama==0.4.4; platform_system == 'Windows'

configparser==5.2.0

flask==2.1.2

google-api-core[grpc]==2.7.3; python_version >= '3.6'

google-auth==2.6.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'

google-cloud-appengine-logging==1.1.1; python_version >= '3.6'

google-cloud-audit-log==0.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

google-cloud-bigquery-storage==2.13.1; python_version >= '3.6'

google-cloud-bigquery==3.1.0

google-cloud-core==2.3.0; python_version >= '3.6'

google-cloud-logging==3.0.0

google-cloud-pubsub==2.12.1

google-crc32c==1.3.0; python_version >= '3.6'

google-resumable-media==2.3.2; python_version >= '3.6'

googleapis-common-protos[grpc]==1.56.0; python_version >= '3.6'

grpc-google-iam-v1==0.12.4; python_version >= '3.6'

grpcio-status==1.46.1

grpcio==1.46.1

idna==3.3; python_version >= '3'

importlib-metadata==4.11.3; python_version < '3.10'

itsdangerous==2.1.2; python_version >= '3.7'

jinja2==3.0.3

jinjasql==0.1.8

markupsafe==2.1.1; python_version >= '3.7'

numpy==1.22.3; python_version < '3.10' and platform_machine != 'aarch64' and platform_machine != 'arm64'

packaging==21.3; python_version >= '3.6'

pandas==1.4.2

proto-plus==1.20.3; python_version >= '3.6'

protobuf==3.20.1; python_version >= '3.7'

psycopg2==2.9.3

pyarrow==8.0.0

pyasn1-modules==0.2.8

pyasn1==0.4.8

pyparsing==3.0.9; python_full_version >= '3.6.8'

python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

pytz==2022.1

requests==2.27.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'

rsa==4.8; python_version >= '3.6'

six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

urllib3==1.26.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'

werkzeug==2.1.2; python_version >= '3.7'

zipp==3.8.0; python_version >= '3.7'

